caca

+++ includes/my_float.h
@@ -6,7 +6,7 @@
 /*   By: mrodrigu <mrodrigu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/19 18:41:29 by mrodrigu          #+#    #+#             */
-/*   Updated: 2018/02/07 07:03:13 by mrodrigu         ###   ########.fr       */
+/*   Updated: 2018/02/09 07:10:01 by jagarcia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ t_myfloat	ft_multiply(t_myfloat x, t_myfloat y);
 t_myfloat	ft_dtomyd(unsigned long int l);
 int			ft_choose_power(int e, int alpha);
 double		ft_sige(double x);
-char		*ft_round(char *str, size_t pos);
+char		*ft_round(char *str, int pos);
 char		*ft_putthepoint(char *str, size_t pos);
 
 #endif
